+++ q-willow.php
@@ -13,7 +13,7 @@
  * Plugin Name:     Q Willow
  * Plugin URI:      https://www.qstudio.us
  * Description:     Willow is a Simple, logic-less, procedural semantic template engine 
- * Version:         1.3.2
+ * Version:         1.3.3
  * Author:          Q Studio
  * Author URI:      https://www.qstudio.us
  * License:         GPL
@@ -244,42 +244,55 @@ if ( ! class_exists( 'q_willow' ) ) {
 			$filter		= null, // array that holds all collected filters to be processed ##
 
 			// list of allowed filters ##
-			// used in tags like [ escape:html, format:uppercase ] can be chained ##
+			// used in tags like [ esc_html, strtoupper ] can be chained ##
 			$filters 	= [
-				'escape' 				=> [
-					'html' 				=> 'esc_html',
-					'attr' 				=> 'esc_attr',
-					'url' 				=> 'esc_url',
-					'urlraw' 			=> 'urlraw',
-					'urlencode' 		=> 'urlencode',
-					'js' 				=> 'esc_js',
-					'textarea' 			=> 'esc_textarea'
-				],
-				'format'				=> [
-					'uppercase' 		=> 'strtoupper',
-					'lowercase' 		=> 'strtolower',
-					'strip_tags' 		=> 'strip_tags',
-					'nl2br' 			=> 'nl2br',
-					'wpautop'			=> 'wpautop',
-					'intval'			=> 'intval',
-					'absint'			=> 'absint',
-				],
-				'sanitize'				=> [
-					'title_with_dashes' => 'sanitize_title_with_dashes',
-					'title' 			=> 'sanitize_title',
-					'email' 			=> 'sanitize_email',
-					'key' 				=> 'sanitize_key',
-					'file_name'			=> 'sanitize_file_name',
-					'html_class'		=> 'sanitize_html_class',
-					'text_field'		=> 'sanitize_text_field',
-					'text_field'		=> 'sanitize_text_field',
-					'user'				=> 'sanitize_user',
-				]
+				// escape ##
+				'esc_html',
+				'esc_attr',
+				'esc_url',
+				'urlraw',
+				'urlencode',
+				'esc_js',
+				'esc_textarea',
+
+				// format ##
+				'strtoupper',
+				'strtolower',
+				'strip_tags',
+				'nl2br',
+				'wpautop',
+				'intval',
+				'absint',
+
+				// sanitize ##
+				'sanitize_title_with_dashes',
+				'sanitize_title',
+				'sanitize_email',
+				'sanitize_key',
+				'sanitize_file_name',
+				'sanitize_html_class',
+				'sanitize_text_field',
+				'sanitize_text_field',
+				'sanitize_user'
 			], 
 
 			// load filters once, with callback filter to allow for changes - this property tracks the load status ##
-			$filters_filtered = false
+			$filters_filtered = false,
 			// $filter_hash
+
+			// allowed flags - currently not referenced or filtered, but might be in future release ##
+			$flags 	= [
+				
+				'buffer', // php_function, willow ##
+				'return', // php_function ##
+				'array', // arguments ##
+				'html', // comments ##
+				'php', // comments ##
+
+			]
+
+			// load flags once, with callback filter to allow for changes - this property tracks the load status ##
+			// $flags_filtered = false
 
 		;